Update BUILD.md
"move the preexisting NorthstarLauncher.exe and Northstar.dll" this confused me because why would the original .exe be named Northstarlauncher? So I changed it to the real names. Plus please include something about Titanfall2_trial.exe because are we suposed to keep it for some reason or do we just back it up like the rest?